requirements.txt
youtube-transcript-api
langchain-community
langchain-openai
faiss-cpu
tiktoken
python-dotenv
euriai
euriai[langchain]
streamlit